requirements.txt
alembic==1.13.1
astroid==3.1.0
async-timeout==4.0.3
attrs==23.2.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
black==24.4.0
blinker==1.7.0
cachelib==0.13.0
certifi==2024.2.2
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
cssselect==1.2.0
DateTime==5.5
Deprecated==1.2.14
dill==0.3.8
distlib==0.3.8
filelock==3.13.4
flake8==7.0.0
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-Migrate==4.0.7
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
gunicorn==22.0.0
identify==2.5.36
idna==3.7
importlib_metadata==7.1.0
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.3
langdetect==1.0.9
Mako==1.3.3
MarkupSafe==2.1.5
mccabe==0.7.0
msgspec==0.18.6
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==24.0
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.1
pluggy==1.5.0
pre-commit==3.7.0
psycopg2-binary==2.9.9
py==1.11.0
pyahocorasick==2.1.0
pycodestyle==2.11.1
pydot==2.0.0
PyExecJS==1.5.1
pyflakes==3.2.0
pylint==3.1.0
pyparsing==3.1.2
pytest==8.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.4
requests==2.31.0
setuptools==69.5.1
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.29
sqlalchemy_schemadisplay==2.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.4
typing_extensions==4.11.0
urllib3==2.2.1
virtualenv==20.26.0
Werkzeug==3.0.2
wrapt==1.16.0
zipp==3.18.1
zope.interface==6.3
